moved options specs into their own spec file
